Move utils directory into src.
