+++ src/proteus/utils/spider.py
@@ -1,8 +1,8 @@
 # Function and classes used to run SPIDER
 
-from utils.modules_ext import *
-from utils.constants import *
-from utils.helper import *
+from proteus.utils.modules_ext import *
+from proteus.utils.constants import *
+from proteus.utils.helper import *
 
 log = logging.getLogger("PROTEUS")
 class MyJSON( object ):
