+++ src/proteus/utils/surface_gases.py
@@ -1,6 +1,6 @@
-from utils.modules_ext import *
-from utils.constants import *
-from utils.helper import *
+from proteus.utils.modules_ext import *
+from proteus.utils.constants import *
+from proteus.utils.helper import *
 
 log = logging.getLogger("PROTEUS")
