+++ src/proteus/utils/templates/init_spider.opts
