+++ src/proteus/utils/templates/vulcan_offline_template.py
